refactor: game storage에서 island storage 분리
