+++ src/modules/game/game.module.ts
@@ -3,6 +3,7 @@ import { GameIslandService } from 'src/domain/services/game/game-island.service'
 import { GameService } from 'src/domain/services/game/game.service';
 import { ChatMessageModule } from 'src/modules/chat-messages/chat-message.module';
 import { GameStorageModule } from 'src/modules/game/game-storage.module';
+import { IslandStorageModule } from 'src/modules/game/island-storage.module';
 import { IslandJoinComponentModule } from 'src/modules/island-joins/island-join-component.module';
 import { IslandComponentModule } from 'src/modules/islands/island-component.module';
 import { UserComponentModule } from 'src/modules/users/users-component.module';
@@ -12,6 +13,7 @@ import { IslandGateway } from 'src/presentation/gateway/island.gateway';
 @Module({
     imports: [
         GameStorageModule,
+        IslandStorageModule,
         UserComponentModule,
         IslandComponentModule,
         IslandJoinComponentModule,
